+++ tests/test_collect.py
@@ -10,13 +10,12 @@ import pytest
 from _pytask.collect import _find_shortest_uniquely_identifiable_name_for_tasks
 from _pytask.collect import pytask_collect_node
 from _pytask.exceptions import NodeNotCollectedError
-from _pytask.nodes import create_task_name
 from pytask import cli
 from pytask import CollectionOutcome
 from pytask import ExitCode
 from pytask import main
-from pytask import PythonFunctionTask
 from pytask import Session
+from pytask import Task
 
 
 @pytest.mark.end_to_end
@@ -54,7 +53,8 @@ def test_collect_filepathnode_with_unknown_type(tmp_path):
 
     assert session.exit_code == 3
     assert session.collection_reports[0].outcome == CollectionOutcome.FAIL
-    assert isinstance(session.collection_reports[0].exc_info[1], NodeNotCollectedError)
+    exc_info = session.collection_reports[0].exc_info
+    assert isinstance(exc_info[1], NodeNotCollectedError)
 
 
 @pytest.mark.end_to_end
@@ -205,20 +205,20 @@ def test_find_shortest_uniquely_identifiable_names_for_tasks(tmp_path):
     path_identifiable_by_base_name = dir_identifiable_by_base_name.joinpath("t.py")
 
     for base_name in ("base_name_ident_0", "base_name_ident_1"):
-        task_id = create_task_name(path_identifiable_by_base_name, base_name)
-        tasks.append(
-            PythonFunctionTask(base_name, task_id, path_identifiable_by_base_name, None)
+        task = Task(
+            base_name=base_name, path=path_identifiable_by_base_name, function=None
         )
-        expected[task_id] = "t.py::" + base_name
+        tasks.append(task)
+        expected[task.name] = "t.py::" + base_name
 
     dir_identifiable_by_module_name = tmp_path.joinpath("identifiable_by_module")
     dir_identifiable_by_module_name.mkdir()
 
     for module in ("t.py", "m.py"):
         module_path = dir_identifiable_by_module_name / module
-        task_id = create_task_name(module_path, "task_a")
-        tasks.append(PythonFunctionTask("task_a", task_id, module_path, None))
-        expected[task_id] = module + "::task_a"
+        task = Task(base_name="task_a", path=module_path, function=None)
+        tasks.append(task)
+        expected[task.name] = module + "::task_a"
 
     dir_identifiable_by_folder = tmp_path / "identifiable_by_folder"
     dir_identifiable_by_folder_a = dir_identifiable_by_folder / "a"
@@ -228,9 +228,9 @@ def test_find_shortest_uniquely_identifiable_names_for_tasks(tmp_path):
 
     for base_path in (dir_identifiable_by_folder_a, dir_identifiable_by_folder_b):
         module_path = base_path / "t.py"
-        task_id = create_task_name(module_path, "task_t")
-        tasks.append(PythonFunctionTask("task_t", task_id, module_path, None))
-        expected[task_id] = base_path.name + "/t.py::task_t"
+        task = Task(base_name="task_t", path=module_path, function=None)
+        tasks.append(task)
+        expected[task.name] = base_path.name + "/t.py::task_t"
 
     result = _find_shortest_uniquely_identifiable_name_for_tasks(tasks)
     assert result == expected