+++ tests/test_console.py
@@ -11,10 +11,9 @@ from _pytask.console import create_url_style_for_path
 from _pytask.console import create_url_style_for_task
 from _pytask.console import format_task_id
 from _pytask.console import render_to_string
-from _pytask.nodes import create_task_name
 from pytask import CollectionOutcome
 from pytask import console
-from pytask import PythonFunctionTask
+from pytask import Task
 from pytask import TaskOutcome
 from rich.console import Console
 from rich.style import Style
@@ -168,9 +167,7 @@ def test_format_task_id(
 ):
     path = _THIS_FILE
 
-    task = PythonFunctionTask(
-        base_name, create_task_name(path, base_name), path, task_func
-    )
+    task = Task(base_name=base_name, path=path, function=task_func)
     if short_name is not None:
         task.short_name = short_name