+++ tests/test_dag.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
 from contextlib import ExitStack as does_not_raise  # noqa: N813
+from pathlib import Path
 
-import attr
 import networkx as nx
 import pytest
 from _pytask.dag import _extract_priorities_from_tasks
@@ -11,37 +11,18 @@ from _pytask.dag import node_and_neighbors
 from _pytask.dag import task_and_descending_tasks
 from _pytask.dag import TopologicalSorter
 from pytask import Mark
-from pytask import MetaTask
-
-
-@attr.s
-class _DummyTask(MetaTask):
-    name = attr.ib(type=str, converter=str)
-    markers = attr.ib(factory=list)
-    path = attr.ib(default=None)
-    base_name = ""
-    short_name = attr.ib(init=False)
-
-    def __attrs_post_init__(self):
-        self.short_name = self.name
-
-    def execute(self):
-        ...
-
-    def state(self):
-        ...
-
-    def add_report_section(self):
-        ...
+from pytask import Task
 
 
 @pytest.fixture()
 def dag():
     dag = nx.DiGraph()
     for i in range(4):
-        dag.add_node(str(i), task=_DummyTask(i))
-        dag.add_node(str(i + 1), task=_DummyTask(i + 1))
-        dag.add_edge(str(i), str(i + 1))
+        dag.add_node(f".::{i}", task=Task(base_name=str(i), path=Path(), function=None))
+        dag.add_node(
+            f".::{i + 1}", task=Task(base_name=str(i + 1), path=Path(), function=None)
+        )
+        dag.add_edge(f".::{i}", f".::{i + 1}")
 
     return dag
 
@@ -49,28 +30,28 @@ def dag():
 @pytest.mark.unit
 def test_sort_tasks_topologically(dag):
     topo_ordering = list(TopologicalSorter.from_dag(dag).static_order())
-    assert topo_ordering == [str(i) for i in range(5)]
+    assert topo_ordering == [f".::{i}" for i in range(5)]
 
 
 @pytest.mark.unit
 def test_descending_tasks(dag):
     for i in range(5):
-        descendants = sorted(descending_tasks(str(i), dag))
-        assert descendants == [str(i) for i in range(i + 1, 5)]
+        descendants = sorted(descending_tasks(f".::{i}", dag))
+        assert descendants == [f".::{i}" for i in range(i + 1, 5)]
 
 
 @pytest.mark.unit
 def test_task_and_descending_tasks(dag):
     for i in range(5):
-        descendants = sorted(task_and_descending_tasks(str(i), dag))
-        assert descendants == [str(i) for i in range(i, 5)]
+        descendants = sorted(task_and_descending_tasks(f".::{i}", dag))
+        assert descendants == [f".::{i}" for i in range(i, 5)]
 
 
 @pytest.mark.unit
 def test_node_and_neighbors(dag):
     for i in range(1, 4):
-        nodes = sorted(node_and_neighbors(dag, str(i)))
-        assert nodes == [str(j) for j in range(i - 1, i + 2)]
+        nodes = sorted(node_and_neighbors(dag, f".::{i}"))
+        assert nodes == [f".::{j}" for j in range(i - 1, i + 2)]
 
 
 @pytest.mark.unit
@@ -78,38 +59,68 @@ def test_node_and_neighbors(dag):
     "tasks, expectation, expected",
     [
         pytest.param(
-            [_DummyTask("1", [Mark("try_last", (), {})])],
+            [
+                Task(
+                    base_name="1",
+                    path=Path(),
+                    function=None,
+                    markers=[Mark("try_last", (), {})],
+                )
+            ],
             does_not_raise(),
-            {"1": -1},
+            {".::1": -1},
             id="test try_last",
         ),
         pytest.param(
-            [_DummyTask("1", [Mark("try_first", (), {})])],
+            [
+                Task(
+                    base_name="1",
+                    path=Path(),
+                    function=None,
+                    markers=[Mark("try_first", (), {})],
+                )
+            ],
             does_not_raise(),
-            {"1": 1},
+            {".::1": 1},
             id="test try_first",
         ),
         pytest.param(
-            [_DummyTask("1", [])], does_not_raise(), {"1": 0}, id="test no priority"
+            [Task(base_name="1", path=Path(), function=None, markers=[])],
+            does_not_raise(),
+            {".::1": 0},
+            id="test no priority",
         ),
         pytest.param(
             [
-                _DummyTask(
-                    "1", [Mark("try_first", (), {}), Mark("try_last", (), {})], ""
+                Task(
+                    base_name="1",
+                    path=Path(),
+                    function=None,
+                    markers=[Mark("try_first", (), {}), Mark("try_last", (), {})],
                 )
             ],
             pytest.raises(ValueError, match="'try_first' and 'try_last' cannot be"),
-            {"1": 1},
+            {".::1": 1},
             id="test mixed priorities",
         ),
         pytest.param(
             [
-                _DummyTask("1", [Mark("try_first", (), {})]),
-                _DummyTask("2", []),
-                _DummyTask("3", [Mark("try_last", (), {})]),
+                Task(
+                    base_name="1",
+                    path=Path(),
+                    function=None,
+                    markers=[Mark("try_first", (), {})],
+                ),
+                Task(base_name="2", path=Path(), function=None, markers=[]),
+                Task(
+                    base_name="3",
+                    path=Path(),
+                    function=None,
+                    markers=[Mark("try_last", (), {})],
+                ),
             ],
             does_not_raise(),
-            {"1": 1, "2": 0, "3": -1},
+            {".::1": 1, ".::2": 0, ".::3": -1},
         ),
     ],
 )
@@ -128,7 +139,7 @@ def test_raise_error_for_undirected_graphs(dag):
 
 @pytest.mark.unit
 def test_raise_error_for_cycle_in_graph(dag):
-    dag.add_edge("4", "1")
+    dag.add_edge(".::4", ".::1")
     scheduler = TopologicalSorter.from_dag(dag)
     with pytest.raises(ValueError, match="The DAG contains cycles."):
         scheduler.prepare()