+++ tests/test_skipping.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import textwrap
 from contextlib import ExitStack as does_not_raise  # noqa: N813
+from pathlib import Path
 
 import pytest
 from _pytask.skipping import pytask_execute_task_setup
@@ -12,6 +13,7 @@ from pytask import Mark
 from pytask import Skipped
 from pytask import SkippedAncestorFailed
 from pytask import SkippedUnchanged
+from pytask import Task
 from pytask import TaskOutcome
 
 
@@ -260,10 +262,7 @@ def test_if_skipif_decorator_is_applied_any_condition_matches(tmp_path):
     ],
 )
 def test_pytask_execute_task_setup(marker_name, expectation):
-    class Task:
-        pass
-
-    task = Task()
+    task = Task(base_name="task", path=Path(), function=None)
     kwargs = {"reason": ""} if marker_name == "skip_ancestor_failed" else {}
     task.markers = [Mark(marker_name, (), kwargs)]