+++ tests/test_task.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import textwrap
 
 import pytest
-from _pytask.nodes import create_task_name
 from pytask import cli
 from pytask import ExitCode
 from pytask import main
@@ -29,13 +28,9 @@ def test_task_with_task_decorator(tmp_path, func_name, task_name):
     assert session.exit_code == ExitCode.OK
 
     if task_name:
-        assert session.tasks[0].name == create_task_name(
-            tmp_path.joinpath("task_module.py"), task_name
-        )
+        assert session.tasks[0].name.endswith(f"task_module.py::{task_name}")
     else:
-        assert session.tasks[0].name == create_task_name(
-            tmp_path.joinpath("task_module.py"), func_name
-        )
+        assert session.tasks[0].name.endswith(f"task_module.py::{func_name}")
 
 
 @pytest.mark.end_to_end
@@ -58,20 +53,13 @@ def test_task_with_task_decorator_with_parametrize(tmp_path, func_name, task_nam
 
     assert session.exit_code == ExitCode.OK
 
+    file_name = path_to_module.name
     if task_name:
-        assert session.tasks[0].name == create_task_name(
-            path_to_module, f"{task_name}[out_1.txt]"
-        )
-        assert session.tasks[1].name == create_task_name(
-            path_to_module, f"{task_name}[out_2.txt]"
-        )
+        assert session.tasks[0].name.endswith(f"{file_name}::{task_name}[out_1.txt]")
+        assert session.tasks[1].name.endswith(f"{file_name}::{task_name}[out_2.txt]")
     else:
-        assert session.tasks[0].name == create_task_name(
-            path_to_module, f"{func_name}[out_1.txt]"
-        )
-        assert session.tasks[1].name == create_task_name(
-            path_to_module, f"{func_name}[out_2.txt]"
-        )
+        assert session.tasks[0].name.endswith(f"{file_name}::{func_name}[out_1.txt]")
+        assert session.tasks[1].name.endswith(f"{file_name}::{func_name}[out_2.txt]")
 
 
 @pytest.mark.end_to_end